Write comments and tests
