--- src/registry.rs
@@ -1,62 +1,0 @@
-use std::sync::atomic::AtomicUsize;
-use std::sync::atomic::Ordering::{Relaxed, Release, SeqCst};
-
-use global;
-
-#[derive(Default, Debug)]
-pub struct Registry {
-    /// The least significant bit is set if the thread is currently pinned. The rest of the bits
-    /// encode the current epoch.
-    state: AtomicUsize,
-}
-
-impl Registry {
-    // FIXME(stjepang): Registrys are stored in a linked list because linked lists are fairly easy
-    // to implement in a lock-free manner. However, traversal is rather slow due to cache misses and
-    // data dependencies. We should experiment with other data structures as well.
-
-    #[inline]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    #[inline]
-    pub fn get_state(&self) -> (bool, usize) {
-        let state = self.state.load(Relaxed);
-        ((state & 1) == 1, state & !1)
-    }
-
-    /// Marks the thread as pinned.
-    ///
-    /// Must not be called if the thread is already pinned!
-    #[inline]
-    pub fn set_pinned(&self) {
-        let epoch = global::epoch().load(Relaxed);
-        let state = epoch | 1;
-
-        // Now we must store `state` into `self.state`. It's important that any succeeding loads
-        // don't get reordered with this store. In order words, this thread's epoch must be fully
-        // announced to other threads. Only then it becomes safe to load from the shared memory.
-        if cfg!(any(target_arch = "x86", target_arch = "x86_64")) {
-            // On x86 architectures we have a choice:
-            // 1. `atomic::fence(SeqCst)`, which compiles to a `mfence` instruction.
-            // 2. `compare_and_swap(_, _, SeqCst)`, which compiles to a `lock cmpxchg` instruction.
-            //
-            // Both instructions have the effect of a full barrier, but the second one seems to be
-            // faster in this particular case.
-            let result = self.state.compare_and_swap(0, state, SeqCst);
-            debug_assert_eq!(0, result, "Registry::set_pinned()'s CAS should succeed.");
-        } else {
-            self.state.store(state, Relaxed);
-            ::std::sync::atomic::fence(SeqCst);
-        }
-    }
-
-    /// Marks the thread as unpinned.
-    #[inline]
-    pub fn set_unpinned(&self) {
-        // Clear the last bit.
-        // We don't need to preserve the epoch, so just store the number zero.
-        self.state.store(0, Release);
-    }
-}
